another attempt at fixing VC6 compilation
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_8_BRANCH@56191 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ tests/streams/fileback.cpp
@@ -20,10 +20,11 @@
 
 
 #ifdef __VISUALC6__
-// need this to be able to use CPPUNIT_ASSERT_EQUAL with wxFileOffset objects
+// there is no support for int64 output in VC6 stream classes so output it as
+// long which should work as long as we don't use >2GB files in this test...
 static std::ostream& operator<<(std::ostream& ostr, const wxFileOffset& fo)
 {
-    ostr << fo;
+    ostr << (long)fo;
     return ostr;
 }
 #endif // __VISUALC6__
